requirements.txt
hdx-python-api==5.5.2
geopandas~=0.10.2
pandas~=1.3.4
Shapely~=1.8.0